Fix creators. (#42)
* Load all Parsers.

* Fix depth creator validation.

* Fix keypoints creator validation.

* Fix docstrings.

* Layout info change.

* Tests fix.
