+++ depthai_nodes/ml/messages/creators/depth.py
@@ -11,7 +11,8 @@ def create_depth_message(
 ) -> dai.ImgFrame:
     """Create a DepthAI message for a depth map.
 
-    @param depth_map: A NumPy array representing the depth map with shape (HW).
+    @param depth_map: A NumPy array representing the depth map with shape HW or NHW/HWN.
+        Here N stands for batch dimension.
     @type depth_map: np.array
     @param depth_type: A string indicating the type of depth map. It can either be
         'relative' or 'metric'.
@@ -19,27 +20,27 @@ def create_depth_message(
     @return: An ImgFrame object containing the depth information.
     @rtype: dai.ImgFrame
     @raise ValueError: If the depth map is not a NumPy array.
-    @raise ValueError: If the depth map is not 2D.
+    @raise ValueError: If the depth map is not 2D or 3D.
+    @raise ValueError: If the depth map shape is not NHW or HWN.
     @raise ValueError: If the depth type is not 'relative' or 'metric'.
+    @raise NotImplementedError: If the depth type is 'metric'.
     """
 
     if not isinstance(depth_map, np.ndarray):
         raise ValueError(f"Expected numpy array, got {type(depth_map)}.")
 
-    if len(depth_map.shape) != 3:
-        raise ValueError(f"Expected 3D input, got {len(depth_map.shape)}D input.")
-
-    if depth_map.shape[0] == 1:
-        depth_map = depth_map[0, :, :]  # CHW to HW
-    elif depth_map.shape[2] == 1:
-        depth_map = depth_map[:, :, 0]  # HWC to HW
-    else:
-        raise ValueError(
-            f"Unexpected image shape. Expected CHW or HWC, got {depth_map.shape}."
-        )
+    if len(depth_map.shape) == 3:
+        if depth_map.shape[0] == 1:
+            depth_map = depth_map[0, :, :]  # CHW to HW
+        elif depth_map.shape[2] == 1:
+            depth_map = depth_map[:, :, 0]  # HWC to HW
+        else:
+            raise ValueError(
+                f"Unexpected image shape. Expected NHW or HWN, got {depth_map.shape}."
+            )
 
     if len(depth_map.shape) != 2:
-        raise ValueError(f"Expected 2D input, got {len(depth_map.shape)}D input.")
+        raise ValueError(f"Expected 2D or 3D input, got {len(depth_map.shape)}D input.")
 
     if depth_type == "relative":
         data_type = dai.ImgFrame.Type.RAW16
